+++ .github/CHANGELOG.md
@@ -152,6 +152,8 @@ which have been refactored and moved out of `physics.fock`.
 
 ### Bug fixes
 
+* The unitary group optimization of the interferometer have been added. The symplectic matrix that describes an interferometer belongs to the intersection of the orthogonal group and the symplectic group, which is a unitary group. It fixed the issue that the optimization of the interferometer was orthogonal group optimization. [(#173)](https://github.com/XanaduAI/MrMustard/pull/173)
+
 * The `Dgate` and the `Rgate` now correctly parse the case when a single scalar is intended as the same parameter
 of a number of gates in pallel.
  [(#180)](https://github.com/XanaduAI/MrMustard/pull/180)
@@ -178,7 +180,7 @@ a pure dm was converted to ket).
 ### Contributors 
 
 This release contains contributions from (in alphabetical order):
-[Sebastian Duque Mesa](https://github.com/sduquemesa), [Filippo Miatto](https://github.com/ziofil)
+[Sebastian Duque Mesa](https://github.com/sduquemesa), [Filippo Miatto](https://github.com/ziofil), [Yuan Yao](https://github.com/sylviemonet)
 
 ---